refactor: move `WebappReconciler` to `webapp/WebappReconciler` and Separated the `WebappEventSourceInitializer`
